fix redundant case warnings
git-svn-id: https://svn.apache.org/repos/asf/lucene/dev/trunk@1146633 13f79535-47bb-0310-9956-ffa450edef68

+++ lucene/contrib/misc/src/java/org/apache/lucene/index/MultiPassIndexSplitter.java
@@ -188,7 +188,7 @@ public class MultiPassIndexSplitter {
 
     @Override
     public int numDocs() {
-      return (int) liveDocs.cardinality();
+      return liveDocs.cardinality();
     }
 
     /**
